refactor(main): move remaining entry related code
