+++ docs/source/source_code/src/entry_handler.rst
@@ -1,0 +1,5 @@
+entry_handler
+=============
+
+.. doxygenfile:: entry_handler.h
+   :allow-dot-graphs:
