+++ src/confighttp.cpp
@@ -29,10 +29,10 @@
 #include "config.h"
 #include "confighttp.h"
 #include "crypto.h"
+#include "entry_handler.h"
 #include "file_handler.h"
 #include "httpcommon.h"
 #include "logging.h"
-#include "main.h"
 #include "network.h"
 #include "nvhttp.h"
 #include "platform/common.h"
