+++ src/platform/linux/kmsgrab.cpp
@@ -15,9 +15,9 @@
 #include <filesystem>
 
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/platform/common.h"
 #include "src/round_robin.h"
+#include "src/thread_pool.h"
 #include "src/utility.h"
 #include "src/video.h"
