+++ src/platform/linux/misc.cpp
@@ -26,8 +26,8 @@
 #include "graphics.h"
 #include "misc.h"
 #include "src/config.h"
+#include "src/entry_handler.h"
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/platform/common.h"
 #include "vaapi.h"
