+++ src/platform/linux/wlgrab.cpp
@@ -5,7 +5,7 @@
 #include "src/platform/common.h"
 
 #include "src/logging.h"
-#include "src/main.h"
+#include "src/thread_pool.h"
 #include "src/video.h"
 
 #include "cuda.h"
