+++ src/platform/windows/display_base.cpp
@@ -16,9 +16,9 @@ typedef long NTSTATUS;
 #include "misc.h"
 #include "src/config.h"
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/platform/common.h"
 #include "src/stat_trackers.h"
+#include "src/thread_pool.h"
 #include "src/video.h"
 
 namespace platf {
