+++ src/platform/windows/input.cpp
@@ -12,9 +12,11 @@
 #include "keylayout.h"
 #include "misc.h"
 #include "src/config.h"
+#include "src/entry_handler.h"
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/platform/common.h"
+#include "src/task_pool.h"
+#include "src/thread_pool.h"
 
 #ifdef __MINGW32__
 DECLARE_HANDLE(HSYNTHETICPOINTERDEVICE);
