+++ src/platform/windows/misc.cpp
@@ -35,8 +35,8 @@
 #define NTDDI_VERSION NTDDI_WIN10
 #include <Shlwapi.h>
 
+#include "src/entry_handler.h"
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/platform/common.h"
 #include "src/utility.h"
 #include <iterator>
