+++ src/process.cpp
@@ -23,9 +23,9 @@
 #include "config.h"
 #include "crypto.h"
 #include "logging.h"
-#include "main.h"
 #include "platform/common.h"
 #include "system_tray.h"
+#include "thread_pool.h"
 #include "utility.h"
 
 #ifdef _WIN32
